// apps/backend/src/generated/openapi/model/snapshotSaveResponse.ts
/**
 * Lapisla Prover API
 * Lapisla Prover is a theorem proving assistant system with a web-based development environment and registry. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SnapshotMeta } from './snapshotMeta';


export interface SnapshotSaveResponse { 
    result: SnapshotSaveResponse.ResultEnum;
    snapshot: SnapshotMeta;
}
export namespace SnapshotSaveResponse {
    export type ResultEnum = 'newly_saved' | 'already_saved';
    export const ResultEnum = {
        NewlySaved: 'newly_saved' as ResultEnum,
        AlreadySaved: 'already_saved' as ResultEnum
    };
}